Reorganized source and header files into module-specific subdirectories
