+++ Chess/include/pieces/Bishop.h
