+++ Chess/include/pieces/Queen.h
