+++ Chess/src/CMakeLists.txt
@@ -1,13 +1,13 @@
 ﻿target_sources(Chess PRIVATE 
     "${CMAKE_CURRENT_SOURCE_DIR}/main.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/Chess.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/ChessBoard.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/GameManager.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/ChessPiece.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/Rook.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/King.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/Bishop.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/Queen.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/Knight.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/PieceFactory.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/board/ChessBoard.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/board/GameManager.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/pieces/ChessPiece.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/pieces/Rook.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/pieces/King.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/pieces/Bishop.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/pieces/Queen.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/pieces/Knight.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/factories/PieceFactory.cpp"
 )
